feat: menu opened on higher res, added settings for all modules
